Supervised-Cross-Modal-Hashing-Retrieval/DCHMT/requirements.txt
tqdm
scipy
numpy
pillow
matplotlib
sklearn
pytorch==1.9.1

